Задача: Ловим события. Вынес логику в отдельный класс-сервис.
